Updated API version to 1.0.2. Please update all file modules!
Finalized, and added, client:utorrent, parser:utorrent and the accompanying ClientWrapper for utorrent.
Added createRemoteTempFile and createRemoteTempDir, and updated the documentation to basefunctions.sh.
Added metaFile parameter to generic files, and altered all file modules to handle it.
Extended file:local to work with local directories as well.
Updated README documentation for 1.0.2 API version.
Added file:fakedata and file:none, which allow for usage of generated fake data or no file at all.
Added Utils/fakedata/ which contains the utility used to generate fake data, both for the module and for manual usage.
Added a HOWTO to get you started.